pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "causal-bayes-opt"
version = "0.1.0"
description = "Amortized Causal Bayesian Optimization framework"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "jax>=0.4.20",
    "numpy>=1.24.0",
    "pyrsistent>=0.19.0",
    "dm-haiku>=0.0.10",
    "optax>=0.1.7",
    "typing-extensions>=4.0.0",
    "psutil (>=7.0.0,<8.0.0)",
    "pydantic (>=2.11.5,<3.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "matplotlib (>=3.10.3,<4.0.0)",
    "gpy (>=1.13.2,<2.0.0)",
    "emukit (>=0.4.11,<0.5.0)",
    "networkx (>=3.5,<4.0)",
    "statsmodels (>=0.14.4,<0.15.0)",
    "pgmpy (>=0.1.26,<0.2.0)",
    "scikit-learn (>=1.0.0,<2.0.0)",
    "scipy (>=1.14.0,<2.0.0)",
    "wandb (>=0.20.1,<0.21.0)",
    "hydra-core (>=1.3.0,<2.0.0)",
    "debugpy (>=1.8.14,<2.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "ipywidgets (>=8.1.7,<9.0.0)",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "hypothesis>=6.0.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]

test = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "hypothesis>=6.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/causal-bayes-opt"
Repository = "https://github.com/yourusername/causal-bayes-opt"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src", "external", "external/parent_scale"]
addopts = ["-v", "--strict-markers"]

[tool.mypy]
python_version = "3.9"
packages = ["causal_bayes_opt"]
mypy_path = "src"
strict = true

[tool.black]
line-length = 100
target-version = ['py39']
include = '\.pyi?$'

[tool.ruff]
select = ["E", "F", "W", "I"]
line-length = 100
target-version = "py39"


[tool.poetry.group.dev.dependencies]
ruff = "^0.11.12"
pytest = "^8.4.0"
ipykernel = "^6.29.5"

